fix post detail override
